Require oomphinc/composer-installers-extender as a hard dependency.
This is necessary because we need certain JavaScript libraries to be installed at particular paths, which can only be done with the extender. Without this change, we would depend on consumers of Lightning Media to have the extender in their dependencies (or else the JavaScript libraries would be installed in unexpected places, breaking our functionality).

Once https://github.com/composer/installers/pull/419 is merged into composer/installers and released, we can cease to rely on the extender.